+++ packages/cli/src/commands/whoami.ts
@@ -13,7 +13,8 @@ export const commandWhoami = new Command("whoami")
       const account = await client.me();
 
       if (!account) {
-        throw new Error("account not found");
+        const environment = client.options.environment;
+        throw new Error(`Account not found in ${environment} environment. Use \`graphand login\``);
       }
 
       spinner.succeed("Fetched current account successfully");