+++ packages/client/src/lib/ClientAdapter.ts
@@ -25,8 +25,6 @@ import {
   defineFieldsProperties,
   ModelData,
   UpdateObject,
-  assignDatamodel,
-  DataModel,
 } from "@graphand/core";
 import { Client } from "./Client.js";
 import { Subject } from "./Subject.js";
@@ -122,7 +120,7 @@ export class ClientAdapter<T extends typeof Model = typeof Model> extends Adapte
         return obj;
       },
       transform(obj) {
-        if (obj["$_next"]) {
+        if (obj && typeof obj === "object" && "$_next" in obj) {
           return obj["$_next"];
         }
 
@@ -902,13 +900,5 @@ export class ClientAdapter<T extends typeof Model = typeof Model> extends Adapte
     for (const instance of this.#store.values()) {
       defineFieldsProperties(instance);
     }
-  }
-
-  static registerModel(model: typeof Model) {
-    const gdx = this.client.options.gdx;
-    if (model.slug && gdx?.datamodels?.[model.slug]) {
-      assignDatamodel(model, gdx.datamodels[model.slug] as ModelJSON<typeof DataModel>);
-    }
-    super.registerModel(model);
   }
 }